Code review updates
Remove tests sub project, all tests are now in core.
Remove TestStreamingBackend, StreamingTest now has the required abstract
methods.
