+++ akka-http-backend/src/test/scala/com/softwaremill/sttp/akkahttp/AkkaHttpStreamingTest.scala
@@ -2,12 +2,12 @@ package com.softwaremill.sttp.akkahttp
 
 import akka.NotUsed
 import akka.actor.ActorSystem
-import akka.stream.{ActorMaterializer, Materializer}
+import akka.stream.ActorMaterializer
 import akka.stream.scaladsl.Source
 import akka.util.ByteString
 import com.softwaremill.sttp.SttpBackend
 import com.softwaremill.sttp.testing.ConvertToFuture
-import com.softwaremill.sttp.testing.streaming.{StreamingTest, TestStreamingBackend}
+import com.softwaremill.sttp.testing.streaming.StreamingTest
 
 import scala.concurrent.Future
 
@@ -15,15 +15,6 @@ class AkkaHttpStreamingTest extends StreamingTest[Future, Source[ByteString, Any
 
   private implicit val actorSystem: ActorSystem = ActorSystem("sttp-test")
   private implicit val materializer: ActorMaterializer = ActorMaterializer()
-
-  override val testStreamingBackend: TestStreamingBackend[Future, Source[ByteString, Any]] =
-    new AkkaHttpTestStreamingBackend(actorSystem)
-}
-
-class AkkaHttpTestStreamingBackend(
-    actorSystem: ActorSystem
-)(implicit materializer: Materializer)
-    extends TestStreamingBackend[Future, Source[ByteString, Any]] {
 
   override implicit val backend: SttpBackend[Future, Source[ByteString, Any]] =
     AkkaHttpBackend.usingActorSystem(actorSystem)
@@ -36,5 +27,4 @@ class AkkaHttpTestStreamingBackend(
 
   override def bodyConsumer(stream: Source[ByteString, Any]): Future[String] =
     stream.map(_.utf8String).runReduce(_ + _)
-
 }
