--- tests/src/test/resources/textfile.txt
@@ -1,100 +1,0 @@
-- Lorem ipsum dolor sit amet
-- Vivamus sem ipsum.
-- Ut molestie.
-- Donec.
-- Fusce non porta.
-- Nulla ac metus. Morbi mattis.
-- Etiam varius.
-- Nulla.
-- Phasellus id mollis.
-- Suspendisse at.
-- Quisque nec leo velit.
-- Fusce.
-- Maecenas nec tristique senectus et.
-- Integer vestibulum lorem fermentum.
-- Vestibulum consectetuer dolor.
-- Lorem ipsum in.
-- Fusce condimentum auctor scelerisque, wisi.
-- Quisque.
-- Curae.
-- Curae, Nullam.
-- Curae, Integer.
-- Vestibulum dignissim massa. Donec.
-- Pellentesque sed sem.
-- Vivamus est.
-- Maecenas elit sed est.
-- Quisque sed tellus.
-- Cum sociis natoque penatibus et.
-- Fusce aliquam.
-- Donec.
-- Sed elementum, sapien accumsan odio.
-- Nam mattis, magna lectus, tincidunt.
-- Pellentesque scelerisque a, sodales.
-- Sed sed condimentum.
-- Curae, In nonummy. Phasellus adipiscing.
-- Vestibulum quis diam mollis.
-- Sed eros. Duis ipsum.
-- Aenean pellentesque at, mollis tempus.
-- Cras ornare facilisis sodales. Aenean.
-- Cum sociis natoque penatibus.
-- Donec id nulla.
-- Quisque ut sapien.
-- Phasellus purus. Proin ultricies.
-- Aliquam auctor neque. Nunc.
-- Nam nunc fringilla non, vehicula.
-- Morbi molestie, felis ut lobortis.
-- Nulla quis.
-- In.
-- Phasellus laoreet urna.
-- Lorem ipsum.
-- Phasellus.
-- Class aptent taciti sociosqu ad.
-- Sed lacinia.
-- Pellentesque dapibus diam. Duis.
-- Suspendisse est. Curabitur.
-- Fusce condimentum justo.
-- Aenean congue quis, faucibus.
-- Ut pharetra leo. Donec.
-- Fusce.
-- Donec porta.
-- Pellentesque orci.
-- Sed.
-- Quisque rutrum, wisi vulputate wisi.
-- Cum sociis.
-- Cras.
-- Sed eros. Curabitur.
-- Proin in velit wisi, tempor.
-- Quisque eu.
-- Proin.
-- Nam pellentesque sed, imperdiet aliquam.
-- Mauris euismod. Sed euismod orci.
-- Etiam.
-- Donec.
-- Fusce wisi a metus. Proin.
-- Phasellus quis.
-- Donec non imperdiet.
-- Aenean vel bibendum a, laoreet.
-- Fusce non enim. Phasellus vulputate.
-- Donec urna elit, sit.
-- Pellentesque habitant morbi.
-- Nulla ante. Curabitur elit. Donec.
-- Cum sociis natoque penatibus.
-- Maecenas eget leo at.
-- Cum sociis natoque penatibus et.
-- Vivamus lacus.
-- Integer.
-- Curae.
-- Maecenas rhoncus. Morbi.
-- Aenean posuere.
-- Duis.
-- Suspendisse a odio fermentum libero.
-- Nam enim. Fusce enim. In.
-- Maecenas.
-- Lorem ipsum primis.
-- Curabitur ac turpis semper sed.
-- Quisque condimentum. Donec sit.
-- Integer convallis non, posuere.
-- Etiam vulputate, odio.
-- Proin id lorem. Donec quis.
-- Curae, Sed nec augue.
-- Aliquam ut turpis.
